backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "url-markdown-workbench-backend"
version = "1.0.0"
description = "Backend for a local URL-to-Markdown workbench"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.8.0",
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
